Tweak defaults in config, adds readme
Tweaks the default port listing in the peacketbeat charm to cover the most common ports people will attempt to relate the charm to. Packetbeat ignores packets not flagged with a protocol so this makes it nice for first timer deployments.

Additionally added a readme for the charm to the layer